+++ lawa/operators/pdeoperators1d/laplaceoperator1d.tcc
@@ -8,8 +8,8 @@ LaplaceOperator1D<T,Basis>::LaplaceOperator1D(const Basis& _basis)
 
 template <typename T, typename Basis>
 T
-LaplaceOperator1D<T,Basis>::operator()(XType xtype1, int j1, int k1,
-                                       XType xtype2, int j2, int k2) const
+LaplaceOperator1D<T,Basis>::operator()(XType xtype1, int j1, long k1,
+                                       XType xtype2, int j2, long k2) const
 {   
     // v_x * u_x
     return integral(j1, k1, xtype1, 1, j2, k2, xtype2, 1);
