+++ lawa/operators/pdeoperators1d/weightedlaplaceoperator1d.h
@@ -43,8 +43,8 @@ class WeightedLaplaceOperator1D{
         WeightedLaplaceOperator1D(const Basis& _basis, Function<T> weightFct);
 
         T
-        operator()(XType xtype1, int j1, int k1,
-                   XType xtype2, int j2, int k2) const;
+        operator()(XType xtype1, int j1, long k1,
+                   XType xtype2, int j2, long k2) const;
 
         T
         operator()(const Index1D &row_index, const Index1D &col_index) const;
