+++ lawa/operators/pdeoperators1d/weightedpdeoperator1d.tcc
@@ -19,8 +19,8 @@ WeightedPDEOperator1D<T, Basis>::WeightedPDEOperator1D(const Basis& _basis,
 
 template <typename T, typename Basis>
 T
-WeightedPDEOperator1D<T, Basis>::operator()(XType xtype1, int j1, int k1,
-                                              XType xtype2, int j2, int k2) const
+WeightedPDEOperator1D<T, Basis>::operator()(XType xtype1, int j1, long k1,
+                                            XType xtype2, int j2, long k2) const
 {
     // diffusion * v_x *  u_x + convection * v * u_x + reaction * v * u
     return    diffusion_integral(j1, k1, xtype1, 1, j2, k2, xtype2, 1)
