+++ lawa/preconditioners/preconditioners1d/diagonallevelpreconditioner1d.tcc
@@ -4,7 +4,7 @@ namespace lawa {
 
 template <typename T>
 T
-DiagonalLevelPreconditioner1D<T>::operator()(XType /*xtype*/, int j, int /*k*/) const
+DiagonalLevelPreconditioner1D<T>::operator()(XType /*xtype*/, int j, long /*k*/) const
 {
     return 1./std::sqrt(1+pow2i<T>(2*j));
 }
