+++ lawa/preconditioners/preconditioners1d/diagonalmatrixpreconditioner1d.h
@@ -33,7 +33,7 @@ class DiagonalMatrixPreconditioner1D
         DiagonalMatrixPreconditioner1D(const BilinearForm &a);
 
         T
-        operator()(XType xtype1, int j1, int k1) const;
+        operator()(XType xtype1, int j1, long k1) const;
 
         T
         operator()(const Index1D &index) const;
