+++ lawa/preconditioners/preconditioners2d/diagonalmatrixpreconditioner2d.tcc
@@ -10,8 +10,8 @@ DiagonalMatrixPreconditioner2D<T,Basis2D,BilinearForm>::DiagonalMatrixPreconditi
 
 template <typename T, typename Basis2D, typename BilinearForm>
 T
-DiagonalMatrixPreconditioner2D<T,Basis2D,BilinearForm>::operator()(XType xtype1, int j1, int k1,
-                                                                   XType xtype2, int j2, int k2) const
+DiagonalMatrixPreconditioner2D<T,Basis2D,BilinearForm>::operator()(XType xtype1, int j1, long k1,
+                                                                   XType xtype2, int j2, long k2) const
 {
     return 1./std::sqrt(fabs(_a(xtype1,j1,k1,xtype2,j2,k2,
                                 xtype1,j1,k1,xtype2,j2,k2)));
