+++ lawa/preconditioners/preconditioners3d/diagonalmatrixpreconditioner3d.h
@@ -35,9 +35,9 @@ class DiagonalMatrixPreconditioner3D
         DiagonalMatrixPreconditioner3D(const BilinearForm &a);
 
         T
-        operator()(XType xtype1, int j1, int k1, 
-                   XType xtype2, int j2, int k2,
-                   XType xtype3, int j3, int k3) const;
+        operator()(XType xtype1, int j1, long k1,
+                   XType xtype2, int j2, long k2,
+                   XType xtype3, int j3, long k3) const;
 
         T
         operator()(const Index3D &index) const;
