+++ lawa/preconditioners/preconditioners3d/diagonalmatrixpreconditioner3d.tcc
@@ -10,9 +10,9 @@ DiagonalMatrixPreconditioner3D<T,Basis3D,BilinearForm>::DiagonalMatrixPreconditi
 
 template <typename T, typename Basis3D, typename BilinearForm>
 T
-DiagonalMatrixPreconditioner3D<T,Basis3D,BilinearForm>::operator()(XType xtype1, int j1, int k1,
-                                                                   XType xtype2, int j2, int k2,
-                                                                   XType xtype3, int j3, int k3) const
+DiagonalMatrixPreconditioner3D<T,Basis3D,BilinearForm>::operator()(XType xtype1, int j1, long k1,
+                                                                   XType xtype2, int j2, long k2,
+                                                                   XType xtype3, int j3, long k3) const
 {
     return 1./std::sqrt(fabs(_a(xtype1,j1,k1, xtype2,j2,k2, xtype3,j3,k3,
                                 xtype1,j1,k1, xtype2,j2,k2, xtype3,j3,k3)));
