+++ lawa/preconditioners/spacetimepreconditioners/rightnormpreconditioner2d_c.h
@@ -36,8 +36,8 @@ class RightNormPreconditioner2D_c
         RightNormPreconditioner2D_c(const Basis2D &basis, T s=2.);  //s=2: A: H^1 -> H^{-1}
 
         T
-        operator()(XType xtype1, int j1, int k1,
-                   XType xtype2, int j2, int k2) const;
+        operator()(XType xtype1, int j1, long k1,
+                   XType xtype2, int j2, long k2) const;
 
         T
         operator()(const Index2D &index) const;
