+++ lawa/righthandsides/smoothrhswithalignedsing2d.h
@@ -33,8 +33,8 @@ struct SmoothRHSWithAlignedSing2D
                                int order, unsigned short _derivx=0, unsigned short _derivy=0);
 
     T
-    operator()(XType xtype_x, int j_x, int k_x,
-               XType xtype_y, int j_y, int k_y) const;
+    operator()(XType xtype_x, int j_x, long k_x,
+               XType xtype_y, int j_y, long k_y) const;
 
     T
     operator()(const Index2D &index) const;
